requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arxiv==2.2.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
biogeme==3.2.14
biogeme-optimization==0.0.10
bitsandbytes==0.42.0
branca==0.8.1
build==1.2.2.post1
ca-certificates==2025.4.26
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cythonbiogeme==1.0.4
dataclasses-json==0.6.7
datascience==0.17.6
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
distro==1.9.0
docstring-parser==0.16
durationpy==0.9
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.9
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
firecrawl-py==2.1.2
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
krb5==1.21.3
kubernetes==32.0.1
langchain==0.3.23
langchain-chroma==0.2.3
langchain-community==0.3.21
langchain-core==0.3.72
langchain-google==0.1.1
langchain-google-community==2.0.7
langchain-google-firestore==0.5.0
langchain-google-genai==2.1.9
langchain-google-vertexai==2.0.20
langchain-text-splitters==0.3.8
langsmith==0.4.13
levenshtein==0.27.1
libcxx==20.1.4
libedit==3.1.20191231
libffi==3.4.4
libsodium==1.0.20
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.34.1
ncurses==6.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openssl==3.5.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
playwright==1.51.0
plotly==6.0.1
posthog==3.25.0
prompt-toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybind11==3.0.1
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
pyee==12.1.1
pygments==2.19.1
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0
python==3.11.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-levenshtein==0.27.1
python_abi==3.11
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
rapidfuzz==3.13.0
readline==8.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==5.1.1
setuptools==75.8.0
sgmllib3k==1.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sqlite==3.45.3
stack_data==0.6.3
starlette==0.45.3
sympy==1.13.3
tavily-python==0.7.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tk==8.6.14
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
typer==0.15.2
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
xyzservices==2025.1.0
xz==5.6.4
yarl==1.19.0
zeromq==4.3.5
zipp==3.21.0
zlib==1.2.13
zstandard==0.23.0